requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.7
Django==2.2.12
django-grappelli==2.14.1
djangorestframework==3.11.0
djongo==1.2.38
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
PyJWT==1.7.1
